[refactor] Remove unused code
