--- exist-core/src/main/java/org/exist/interpreter/Compiled.java
@@ -1,62 +1,0 @@
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.interpreter;
-
-import org.exist.source.Source;
-import org.exist.xquery.XPathException;
-import org.exist.xquery.XQueryContext;
-import org.exist.xquery.value.Sequence;
-
-public interface Compiled extends IPathExpr {
-
-    public void reset();
-    
-    /**
-     * @return the {@link Context} used to create this query
-     */
-    public XQueryContext getContext();
-    
-    public void setContext(Context context);
-    
-    /**
-     * Execute the compiled query, optionally using the specified
-     * sequence as context.
-     *
-     * @param contextSequence the context sequence.
-     *
-     * @throws XPathException if an error occurs during evaluation.
-     */
-    public Sequence eval(Sequence contextSequence) throws XPathException;
-
-    /**
-     * Is the compiled expression still valid? Returns false if, for example,
-     * the source code of one of the imported modules has changed.
-     */
-    public boolean isValid();
-
-    /**
-     * Gets the source of this query.
-     *
-     * @return This query's source
-     */
-    public Source getSource();
-}
