--- exist-core/src/main/java/org/exist/interpreter/IPragma.java
@@ -1,50 +1,0 @@
-/*
- * eXist-db Open Source Native XML Database
- * Copyright (C) 2001 The eXist-db Authors
- *
- * info@exist-db.org
- * http://www.exist-db.org
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-package org.exist.interpreter;
-
-import org.exist.dom.QName;
-import org.exist.xquery.AnalyzeContextInfo;
-import org.exist.xquery.Expression;
-import org.exist.xquery.XPathException;
-import org.exist.xquery.value.Item;
-import org.exist.xquery.value.Sequence;
-
-public interface IPragma {
-
-	public void analyze(AnalyzeContextInfo contextInfo) throws XPathException;
-
-	public Sequence eval(Sequence contextSequence, Item contextItem) throws XPathException;
-
-	public void before(Context context) throws XPathException;
-
-	public void before(Context context, Expression expression) throws XPathException;
-
-	public void after(Context context) throws XPathException;
-
-	public void after(Context context, Expression expression) throws XPathException;
-
-	public void resetState(boolean postOptimization);
-
-	public String getContents();
-
-	public QName getQName();
-}
